+++ addons/yat/src/enums/MethodValidationResult.cs
@@ -1,0 +1,8 @@
+namespace YAT.Enums;
+
+public enum MethodValidationResult
+{
+	Success,
+	InvalidInstance,
+	InvalidMethod,
+}