+++ addons/yat/src/helpers/Messages.cs
@@ -12,6 +12,8 @@ public static class Messages
 
 	public static string InvalidNodePath(string path) => $"{path} is not a valid node path.";
 	public static string InvalidArgument(string command, string argument, string expected) => $"{command} expected {argument} to be an {expected}.";
+	public static string InvalidMethod(string method) => $"{method} is not a valid method.";
 
 	public static string NodeHasNoChildren(string path) => $"{path} has no children.";
+	public static string DisposedNode => "The node has been disposed.";
 }