+++ CHANGELOG.md
@@ -27,6 +27,7 @@ All notable changes to this project will be documented in this file.
 -   Renamed NumericHelper to Numeric & StorageHelper to Storage.
 -   Improved exception handling in TryConvert method in Numeric class.
 -   Refreshed example scenes.
+-   Testing on Godot SDK 4.3.0-dev.3.
 
 ### Fixed